Move deleteUnusedLogAction logic to RawLogDao class.
