+++ plugins/PrivacyManager/LogDataPurger.php
@@ -22,17 +22,10 @@ use Piwik\Piwik;
  */
 class LogDataPurger
 {
-    const TEMP_TABLE_NAME = 'tmp_log_actions_to_keep';
-
     /**
      * The max set of rows each table scan select should query at one time.
      */
     public static $selectSegmentSize = 100000;
-
-    /**
-     * @param DimensionMetadataProvider
-     */
-    private $dimensionMetadataProvider;
 
     /**
      * TODO
@@ -58,9 +51,8 @@ class LogDataPurger
     /**
      * Constructor.
      */
-    public function __construct(DimensionMetadataProvider $dimensionMetadataProvider, LogPurger $logPurger, RawLogDao $rawLogDao)
+    public function __construct(LogPurger $logPurger, RawLogDao $rawLogDao)
     {
-        $this->dimensionMetadataProvider = $dimensionMetadataProvider;
         $this->logPurger = $logPurger;
         $this->rawLogDao = $rawLogDao;
     }
@@ -96,7 +88,7 @@ class LogDataPurger
 
         // delete unused actions from the log_action table (but only if we can lock tables)
         if (Db::isLockPrivilegeGranted()) {
-            $this->purgeUnusedLogActions(); // TODO: move actual code to DAO/service in core
+            $this->rawLogDao->deleteUnusedLogActions();
         } else {
             $logMessage = get_class($this) . ": LOCK TABLES privilege not granted; skipping unused actions purge";
             Log::warning($logMessage);
@@ -143,29 +135,6 @@ class LogDataPurger
     }
 
     /**
-     * Safely delete all unused log_action rows.
-     */
-    private function purgeUnusedLogActions()
-    {
-        $this->createTempTable();
-
-        // get current max ID in log tables w/ idaction references.
-        $maxIds = $this->getMaxIdsInLogTables();
-
-        // do large insert (inserting everything before maxIds) w/o locking tables...
-        $this->insertActionsToKeep($maxIds, $deleteOlderThanMax = true);
-
-        // ... then do small insert w/ locked tables to minimize the amount of time tables are locked.
-        $this->lockLogTables();
-        $this->insertActionsToKeep($maxIds, $deleteOlderThanMax = false);
-
-        // delete before unlocking tables so there's no chance a new log row that references an
-        // unused action will be inserted.
-        $this->deleteUnusedActions();
-        Db::unlockAllTables();
-    }
-
-    /**
      * get highest idVisit to delete rows from
      * @return string
      */
@@ -196,101 +165,6 @@ class LogDataPurger
     {
         $sql = "SELECT COUNT(*) FROM $table WHERE idvisit <= ?";
         return (int) Db::fetchOne($sql, array($maxIdVisit));
-    }
-
-    private function createTempTable()
-    {
-        $sql = "CREATE TEMPORARY TABLE " . Common::prefixTable(self::TEMP_TABLE_NAME) . " (
-					idaction INT(11),
-					PRIMARY KEY (idaction)
-				)";
-        Db::query($sql);
-    }
-
-    private function getMaxIdsInLogTables()
-    {
-        $tables = array('log_conversion', 'log_link_visit_action', 'log_visit', 'log_conversion_item');
-        $idColumns = $this->getTableIdColumns();
-
-        $result = array();
-        foreach ($tables as $table) {
-            $idCol = $idColumns[$table];
-            $result[$table] = Db::fetchOne("SELECT MAX($idCol) FROM " . Common::prefixTable($table));
-        }
-
-        return $result;
-    }
-
-    private function insertActionsToKeep($maxIds, $olderThan = true)
-    {
-        $tempTableName = Common::prefixTable(self::TEMP_TABLE_NAME);
-
-        $idColumns = $this->getTableIdColumns();
-        $idActionColumnsByTable = $this->dimensionMetadataProvider->getActionReferenceColumnsByTable();
-        foreach ($idActionColumnsByTable as $table => $columns) {
-            $idCol = $idColumns[$table];
-
-            foreach ($columns as $col) {
-                $select = "SELECT $col FROM " . Common::prefixTable($table) . " WHERE $idCol >= ? AND $idCol < ?";
-                $sql = "INSERT IGNORE INTO $tempTableName $select";
-
-                if ($olderThan) {
-                    $start = 0;
-                    $finish = $maxIds[$table];
-                } else {
-                    $start = $maxIds[$table];
-                    $finish = Db::fetchOne("SELECT MAX($idCol) FROM " . Common::prefixTable($table));
-                }
-
-                Db::segmentedQuery($sql, $start, $finish, self::$selectSegmentSize);
-            }
-        }
-
-        // allow code to be executed after data is inserted. for concurrency testing purposes.
-        if ($olderThan) {
-            /**
-             * @ignore
-             */
-            Piwik::postEvent("LogDataPurger.ActionsToKeepInserted.olderThan");
-        } else {
-            /**
-             * @ignore
-             */
-            Piwik::postEvent("LogDataPurger.ActionsToKeepInserted.newerThan");
-        }
-    }
-
-    private function lockLogTables()
-    {
-        Db::lockTables(
-            $readLocks = Common::prefixTables('log_conversion',
-                'log_link_visit_action',
-                'log_visit',
-                'log_conversion_item'),
-            $writeLocks = Common::prefixTables('log_action')
-        );
-    }
-
-    private function deleteUnusedActions()
-    {
-        list($logActionTable, $tempTableName) = Common::prefixTables("log_action", self::TEMP_TABLE_NAME);
-
-        $deleteSql = "DELETE LOW_PRIORITY QUICK IGNORE $logActionTable
-						FROM $logActionTable
-				   LEFT JOIN $tempTableName tmp ON tmp.idaction = $logActionTable.idaction
-					   WHERE tmp.idaction IS NULL";
-
-        Db::query($deleteSql);
-    }
-
-    private function getTableIdColumns()
-    {
-        return array(
-            'log_link_visit_action' => 'idlink_va',
-            'log_conversion'        => 'idvisit',
-            'log_visit'             => 'idvisit',
-            'log_conversion_item'   => 'idvisit'
-        );
     }
 
     // let's hardcode, since these are not dynamically created tables
